oldpyrobo/requirements.txt
pyyaml
discord.py
youtube-search-python
Mastodon.py
python-dateutil
black
pycurl
